Add files via upload
Loss: Loss represents how well the model is performing on the training data. It is a measure of the error between the actual outputs and the predicted outputs. The loss function used here is binary cross-entropy since it's a binary classification problem (normal vs. anomaly detection).

Accuracy: Accuracy measures the proportion of correct predictions made by the model on the test data. It is calculated by dividing the number of correct predictions by the total number of predictions made.